- up to 2010.12.04
Changed files:
    GeoIP-db-Country.spec -> 1.17

+++ GeoIP-db-Country.spec
@@ -2,12 +2,12 @@ Summary:	GeoLite Country - Country database for GeoIP
 Summary(pl.UTF-8):	GeoLite Country - baza danych krajów dla GeoIP
 Name:		GeoIP-db-Country
 # Updated every month:
-Version:	2010.11.02
+Version:	2010.12.04
 Release:	1
 License:	OPEN DATA LICENSE (see LICENSE.txt)
 Group:		Applications/Databases
 Source0:	http://geolite.maxmind.com/download/geoip/database/GeoLiteCountry/GeoIP.dat.gz
-# Source0-md5:	b28f1b22ee576da7ec7333f5b8cf7cd5
+# Source0-md5:	42a6b039337ed271ee0c8f40497ceba9
 Source1:	http://www.maxmind.com/download/geoip/database/LICENSE.txt
 # Source1-md5:	a1381bd1aa0a0c91dc31b3f1e847cf4a
 URL:		http://www.maxmind.com/app/geolitecountry
@@ -44,17 +44,21 @@ Informacja licencyjna: ten produkt zawiera dane GeoLite stworzone
 przez MaxWind, dostępne z <http://www.maxwind.com/>.
 
 %prep
-%setup -q -c -T
-cp %{SOURCE0} .
-cp %{SOURCE1} .
+%setup -qcT
+cp -a %{SOURCE0} .
+cp -a %{SOURCE1} .
 
 gunzip GeoIP.dat.gz
+
+ver=$(stat -c '%y' GeoIP.dat | awk '{print $1}' | tr - .)
+if [ "$ver" != %{version} ]; then
+	exit 1
+fi
 
 %install
 rm -rf $RPM_BUILD_ROOT
 install -d $RPM_BUILD_ROOT%{_datadir}/GeoIP
-
-install GeoIP.dat $RPM_BUILD_ROOT%{_datadir}/GeoIP
+cp -a GeoIP.dat $RPM_BUILD_ROOT%{_datadir}/GeoIP
 
 %clean
 rm -rf $RPM_BUILD_ROOT
